Another attempt at enabling fstrim.timer on F32 upgrades (#1811506)
Avoid using %systemd_post as it is guarding the trigger with '$1 -eq 1'
which not true for %triggerpostun (it's 2 for updates in triggerpostun).

Also, while at this, fine-tune the trigger condition to trigger for
package updates from older F32 versions as well -- this makes it much
easier to test this fix, and also fixes fstrim.timer for anyone who
installed from F32 beta media.