Simplify creation of /etc/mtab symlink
It was created unconditionally in %post, so the original contents (if
any) were overwritten. Listing is as %config was just misleading.
Let's simply list the symlink as normal contents in the rpm, so that
rpm can install the symlink for us.

(The scriplet was missing a dependency on coreutils for ln, but let's
not add that if we can drop the whole thing.)

[kzak@redhat.com: - add %changelog entry]